Added support for search in comments

+++ backend/controllers/posts.controller.js
@@ -3,19 +3,24 @@ import {
   createPost,
   deletePost,
   getAllPosts,
+  searchPosts,
   updatePost,
 } from "../models/post.model.js";
 import {
   commentsCollection,
   postsCollection,
 } from "../utils/collections.util.js";
-import { deleteComment, updateComment } from "../models/comment.model.js";
+import {
+  deleteComment,
+  searchComments,
+  updateComment,
+} from "../models/comment.model.js";
 
 const router = express.Router();
 
 router.post("/", async (req, res, next) => {
   try {
-    req.body.userId = req.userId;
+    req.body.user = req.userId;
     const response = await createPost(req.body);
     return res.status(200).json(response);
   } catch (error) {
@@ -25,8 +30,18 @@ router.post("/", async (req, res, next) => {
 
 router.get("/", async (req, res, next) => {
   try {
-    const response = await getAllPosts();
-    return res.status(200).json(response);
+    const searchString = req.query.searchString;
+    if (searchString) {
+      const postsResponse = await searchPosts(searchString);
+      const commentsResponse = await searchComments(searchString);
+      return res.status(200).json({
+        posts: postsResponse,
+        comments: commentsResponse,
+      });
+    } else {
+      const response = await getAllPosts();
+      return res.status(200).json(response);
+    }
   } catch (error) {
     next(error);
   }
@@ -34,9 +49,11 @@ router.get("/", async (req, res, next) => {
 
 router.post("/post/:id/comment", async (req, res, next) => {
   try {
+    const userId = req.userId;
     const comment = await commentsCollection.create({
       content: req.body.content,
       post: req.params.id,
+      userId: userId,
     });
     // console.log("here");
     const postRelated = await postsCollection.findById(req.params.id);
@@ -52,6 +69,7 @@ router.put("/post/:id/comment/:cid", async (req, res, next) => {
   try {
     const postId = req.params.id;
     const commentId = req.params.cid;
+    const userId = req.userId;
     const postRelated = await postsCollection.findById(postId);
     const commentRelated = postRelated.comments.filter(
       (comment) => comment._id == commentId
@@ -60,6 +78,9 @@ router.put("/post/:id/comment/:cid", async (req, res, next) => {
       throw new Error(
         `No comment with ${commentId} exist in this post ${postId}`
       );
+    }
+    if (commentRelated?.user?._id.toString() !== userId) {
+      throw new Error("You cannot update someone else's comment");
     }
     const upatedComment = await updateComment(
       {
@@ -77,7 +98,11 @@ router.delete("/post/:id/comment/:cid", async (req, res, next) => {
   try {
     const postId = req.params.id;
     const commentId = req.params.cid;
+    const userId = req.userId;
     const postRelated = await postsCollection.findById(postId);
+    if (postRelated?.user?._id.toString() !== userId) {
+      throw new Error("You cannot delete comments from someone else's post");
+    }
     const commentIndex = postRelated.comments.findIndex(
       (comment) => comment._id.toString() === commentId
     );
@@ -96,7 +121,11 @@ router.delete("/post/:id/comment/:cid", async (req, res, next) => {
 
 router.put("/:id", async (req, res, next) => {
   try {
-    // const userId = req.userId;
+    const userId = req.userId;
+    const postRelated = await postsCollection.findById(req.params.id);
+    if (postRelated?.user?._id.toString() !== userId) {
+      throw new Error("You cannot update someone else's post");
+    }
     const response = await updatePost({ _id: req.params.id }, req.body);
     return res.status(200).json(response);
   } catch (error) {
@@ -106,6 +135,11 @@ router.put("/:id", async (req, res, next) => {
 
 router.delete("/:id", async (req, res, next) => {
   try {
+    const userId = req.userId;
+    const postRelated = await postsCollection.findById(req.params.id);
+    if (postRelated?.user?._id.toString() !== userId) {
+      throw new Error("You cannot delete someone else's post");
+    }
     const response = await deletePost({ _id: req.params.id });
     if (response) {
       return res
